feat: in-place symplectic transformations
Change the functions in `operations.jl` to mutating one, i.e. instead
of returning a new `GaussianState` object the one passed as first
argument is mutated. Usually we want to apply a sequence of operations
to an initial state, without defining a new one each time, so this
syntax is more natural.
The main implementation of the operations is now in the mutating
functions, the old non-mutating functions remain but they just create a
copy of the Gaussian state and act on the copy.