refactor: permute without Permutations
Use the simpler syntax `v[p]` or `m[p, p]` to permute vectors and
matrices given a permutation `p` instead of creating a matrix with the
aid of the Permutations package and then multiplying `v` and `m`
accordingly. This method should be faster since it doesn't require
creating new matrices or any matrix multiplication.
The Permutations package is not needed anymore and has been removed.