fix: make `lossybeamsplitter!` truly mutating
Fix the `lossybeamsplitter!` function so that it truly mutates the
Gaussian state passed as the first argument. In this case there isn't a
mutating `partialtrace` function (yet) so contrary to the other
operations the non-mutating `lossybeamsplitter` is the main function,
while `lossybeamsplitter!` simply replaces the original Gaussian state
with the new one created by `lossybeamsplitter`.