fix: Euler decomposition
The Euler decomposition returned a clearly wrong output (`L` and `R`
matrices were not unitary) with some big matrices. The actual source
of the error is still unclear. The `_euler_xxpp` function was replaced
by the analogue from the `BosonSupremacy` repository, which seems to
work correctly in those cases.

The new function relies upon the SVD decomposition, for which we don't have
a method that works with `BigFloat` numbers (yet), so, for the time
being, the `BigFloat` test for the Euler decomposition function has been
removed.